charms/kfp-persistence/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile ./requirements.in
#
anyio==3.7.1
    # via httpcore
attrs==23.1.0
    # via jsonschema
certifi==2023.5.7
    # via
    #   httpcore
    #   httpx
    #   requests
charmed-kubeflow-chisme @ git+https://github.com/canonical/charmed-kubeflow-chisme.git@add-base-charm
    # via -r ./requirements.in
charset-normalizer==3.2.0
    # via requests
deepdiff==6.2.1
    # via charmed-kubeflow-chisme
dict==2020.12.3
    # via -r ./requirements.in
exceptiongroup==1.1.2
    # via anyio
h11==0.14.0
    # via httpcore
httpcore==0.17.3
    # via httpx
httpx==0.24.1
    # via lightkube
idna==3.4
    # via
    #   anyio
    #   httpx
    #   requests
importlib-resources==6.0.0
    # via jsonschema
jinja2==3.1.2
    # via charmed-kubeflow-chisme
jsonschema==4.17.3
    # via
    #   -r ./requirements.in
    #   serialized-data-interface
lightkube==0.14.0
    # via charmed-kubeflow-chisme
lightkube-models==1.27.1.4
    # via lightkube
markupsafe==2.1.3
    # via jinja2
ops==2.4.1
    # via
    #   -r ./requirements.in
    #   charmed-kubeflow-chisme
    #   serialized-data-interface
ordered-set==4.1.0
    # via deepdiff
pkgutil-resolve-name==1.3.10
    # via jsonschema
pyrsistent==0.19.3
    # via jsonschema
pyyaml==6.0.1
    # via
    #   lightkube
    #   ops
    #   serialized-data-interface
requests==2.31.0
    # via serialized-data-interface
ruamel-yaml==0.17.32
    # via charmed-kubeflow-chisme
ruamel-yaml-clib==0.2.7
    # via ruamel-yaml
serialized-data-interface==0.6.0
    # via -r ./requirements.in
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
tenacity==8.2.2
    # via charmed-kubeflow-chisme
urllib3==2.0.4
    # via requests
values==2020.12.3
    # via dict
websocket-client==1.6.1
    # via ops
zipp==3.16.2
    # via importlib-resources
